making printing of the HTML representation of module handbooks faster
fixes a bug where the materialized view for module atomics would not be refreshed
